requirements.txt
gitpython==2.0.6
prospector==0.12
Flask==0.11.1
watchdog==0.8.3
flask-socketio==2.4
eventlet==0.19.0